added mayor changes and fixed unsecured queries

+++ backend/src/main/java/com/example/service/Impl/GenericServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.persistence.IGenericDAO;
 import com.example.qualifier.FindEntities;
 import com.example.qualifier.FindEntitiesSet;
 import com.example.qualifier.FindEntity;
+import com.example.security.manager.CookieManager;
 import com.example.service.IGenericService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +18,6 @@ public abstract class GenericServiceImpl<T,ID> implements
         IGenericService<T,ID> {
 
     protected abstract IGenericDAO<T, ID> getDAO();
-
     protected abstract String getEntityName();
 
     @FindEntity
@@ -27,15 +27,15 @@ public abstract class GenericServiceImpl<T,ID> implements
         if (id == null) throw new BadRequestException(
                 String.format("%s cannot be found", getEntityName()));
         return getDAO().findById(id).orElseThrow(() -> new NotFoundException(
-                String.format("%s with ID %s not found",
-                        getEntityName(), id.toString())));
+                String.format("%s with ID %s not found", getEntityName(), id)));
     }
 
     @FindEntities
     @Transactional
     @Override
     public List<T> findAllByIds(Collection<ID> ids) {
-        if(ids == null || ids.isEmpty()) throw new BadRequestException(
+        if(ids == null || ids.isEmpty() || ids.stream().anyMatch(Objects::isNull))
+            throw new BadRequestException(
                 String.format("%s(s) cannot be found", getEntityName()));
         if(!existsAllByIds(ids)) throw new NotFoundException(
                 String.format("Some %s(s) cannot be found", getEntityName()));
@@ -60,6 +60,8 @@ public abstract class GenericServiceImpl<T,ID> implements
     @Transactional
     @Override
     public List<T> saveAll(List<T> t) {
+        if(t == null || t.isEmpty()) throw new BadRequestException(
+                String.format("%s cannot be created", getEntityName()));
         return getDAO().saveAll(t);
     }
 
@@ -68,6 +70,7 @@ public abstract class GenericServiceImpl<T,ID> implements
     public T updateById(T t, ID id) {
         if(id == null || t == null) throw new BadRequestException(
                 String.format("%s cannot be updated", getEntityName()));
+        System.out.println("Objeto para actualizar: " + CookieManager.justSerialize(t));
         return Optional.ofNullable(getDAO().updateById(t, id)).orElseThrow(() ->
                 new NotFoundException(String.format("%s with ID %s not found",
                         getEntityName(), id.toString())));
@@ -78,6 +81,8 @@ public abstract class GenericServiceImpl<T,ID> implements
     public void deleteById(ID id) {
         if(id == null) throw new BadRequestException(
                 String.format("%s cannot be deleted", getEntityName()));
+        if(!existsById(id)) throw new BadRequestException(
+                String.format("%s not exists", getEntityName()));
         getDAO().deleteById(id);
     }
 
@@ -90,24 +95,29 @@ public abstract class GenericServiceImpl<T,ID> implements
     @Transactional
     @Override
     public Page<T> findAllPage(Pageable pageable) {
+        if(pageable == null) throw new BadRequestException(
+                String.format("%s(s) Page cannot be created", getEntityName()));
         return getDAO().findAllPage(pageable);
     }
 
     @Transactional
     @Override
     public boolean existsById(ID id) {
+        if (id == null) return false;
         return getDAO().existsById(id);
     }
 
     @Transactional
     @Override
     public boolean existsAllByIds(Collection<ID> ids) {
-        return getDAO().existsAllById(ids);
+        if (ids == null || ids.isEmpty() || ids.stream().anyMatch(Objects::isNull)) return false;
+        return getDAO().existsAllByIds(ids);
     }
 
     @Transactional
     @Override
     public boolean existsByUniqueProperties(T t) {
+        if(t == null) return false;
         return getDAO().existsByUniqueProperties(t);
     }
